+++ labware/Labware definitions/4Ti_0131_Reservoir_12_trough_21000ul/4Ti_0131_Reservoir_12_trough_21000ul.json
@@ -1,0 +1,218 @@
+{
+    "ordering": [
+        [
+            "A1"
+        ],
+        [
+            "A2"
+        ],
+        [
+            "A3"
+        ],
+        [
+            "A4"
+        ],
+        [
+            "A5"
+        ],
+        [
+            "A6"
+        ],
+        [
+            "A7"
+        ],
+        [
+            "A8"
+        ],
+        [
+            "A9"
+        ],
+        [
+            "A10"
+        ],
+        [
+            "A11"
+        ],
+        [
+            "A12"
+        ]
+    ],
+    "brand": {
+        "brand": "4Ti",
+        "brandId": [
+            "4ti-0131"
+        ]
+    },
+    "metadata": {
+        "displayName": "4Ti-0131 Reservoir 12 Trough 21000 µL",
+        "displayCategory": "wellPlate",
+        "displayVolumeUnits": "µL",
+        "tags": []
+    },
+    "dimensions": {
+        "xDimension": 127.2,
+        "yDimension": 85.3,
+        "zDimension": 44.1
+    },
+    "wells": {
+        "A1": {
+            "depth": 39.2,
+            "totalLiquidVolume": 21000,
+            "shape": "rectangular",
+            "xDimension": 9,
+            "yDimension": 71.8,
+            "x": 12.4,
+            "y": 42.6,
+            "z": 4.9
+        },
+        "A2": {
+            "depth": 39.2,
+            "totalLiquidVolume": 21000,
+            "shape": "rectangular",
+            "xDimension": 9,
+            "yDimension": 71.8,
+            "x": 21.4,
+            "y": 42.6,
+            "z": 4.9
+        },
+        "A3": {
+            "depth": 39.2,
+            "totalLiquidVolume": 21000,
+            "shape": "rectangular",
+            "xDimension": 9,
+            "yDimension": 71.8,
+            "x": 30.4,
+            "y": 42.6,
+            "z": 4.9
+        },
+        "A4": {
+            "depth": 39.2,
+            "totalLiquidVolume": 21000,
+            "shape": "rectangular",
+            "xDimension": 9,
+            "yDimension": 71.8,
+            "x": 39.4,
+            "y": 42.6,
+            "z": 4.9
+        },
+        "A5": {
+            "depth": 39.2,
+            "totalLiquidVolume": 21000,
+            "shape": "rectangular",
+            "xDimension": 9,
+            "yDimension": 71.8,
+            "x": 48.4,
+            "y": 42.6,
+            "z": 4.9
+        },
+        "A6": {
+            "depth": 39.2,
+            "totalLiquidVolume": 21000,
+            "shape": "rectangular",
+            "xDimension": 9,
+            "yDimension": 71.8,
+            "x": 57.4,
+            "y": 42.6,
+            "z": 4.9
+        },
+        "A7": {
+            "depth": 39.2,
+            "totalLiquidVolume": 21000,
+            "shape": "rectangular",
+            "xDimension": 9,
+            "yDimension": 71.8,
+            "x": 66.4,
+            "y": 42.6,
+            "z": 4.9
+        },
+        "A8": {
+            "depth": 39.2,
+            "totalLiquidVolume": 21000,
+            "shape": "rectangular",
+            "xDimension": 9,
+            "yDimension": 71.8,
+            "x": 75.4,
+            "y": 42.6,
+            "z": 4.9
+        },
+        "A9": {
+            "depth": 39.2,
+            "totalLiquidVolume": 21000,
+            "shape": "rectangular",
+            "xDimension": 9,
+            "yDimension": 71.8,
+            "x": 84.4,
+            "y": 42.6,
+            "z": 4.9
+        },
+        "A10": {
+            "depth": 39.2,
+            "totalLiquidVolume": 21000,
+            "shape": "rectangular",
+            "xDimension": 9,
+            "yDimension": 71.8,
+            "x": 93.4,
+            "y": 42.6,
+            "z": 4.9
+        },
+        "A11": {
+            "depth": 39.2,
+            "totalLiquidVolume": 21000,
+            "shape": "rectangular",
+            "xDimension": 9,
+            "yDimension": 71.8,
+            "x": 102.4,
+            "y": 42.6,
+            "z": 4.9
+        },
+        "A12": {
+            "depth": 39.2,
+            "totalLiquidVolume": 21000,
+            "shape": "rectangular",
+            "xDimension": 9,
+            "yDimension": 71.8,
+            "x": 111.4,
+            "y": 42.6,
+            "z": 4.9
+        }
+    },
+    "groups": [
+        {
+            "metadata": {
+                "wellBottomShape": "v"
+            },
+            "wells": [
+                "A1",
+                "A2",
+                "A3",
+                "A4",
+                "A5",
+                "A6",
+                "A7",
+                "A8",
+                "A9",
+                "A10",
+                "A11",
+                "A12"
+            ]
+        }
+    ],
+    "parameters": {
+        "format": "irregular",
+        "quirks": [
+            "centerMultichannelOnWells",
+            "touchTipDisabled"
+        ],
+        "isTiprack": false,
+        "isMagneticModuleCompatible": false,
+        "loadName": "4Ti_0131_Reservoir_12_trough_21000ul"
+    },
+    "namespace": "custom_beta",
+    "version": 1,
+    "schemaVersion": 2,
+    "cornerOffsetFromSlot": {
+        "x": 0,
+        "y": 0,
+        "z": 0
+    }
+}
