+++ labware/Labware definitions/E1415-1500_starlabtubes_24_tuberack_1500ul/E1415-1500_starlabtubes_24_tuberack_1500ul.json
@@ -1,0 +1,324 @@
+{
+    "ordering": [
+        [
+            "A1",
+            "B1",
+            "C1",
+            "D1"
+        ],
+        [
+            "A2",
+            "B2",
+            "C2",
+            "D2"
+        ],
+        [
+            "A3",
+            "B3",
+            "C3",
+            "D3"
+        ],
+        [
+            "A4",
+            "B4",
+            "C4",
+            "D4"
+        ],
+        [
+            "A5",
+            "B5",
+            "C5",
+            "D5"
+        ],
+        [
+            "A6",
+            "B6",
+            "C6",
+            "D6"
+        ]
+    ],
+    "brand": {
+        "brand": "Starlab",
+        "brandId": [
+            "E1415-1500"
+        ]
+    },
+    "metadata": {
+        "displayName": "E1415-1500 starlab microcentrifuge tube 24 tuberack_1500ul",
+        "displayCategory": "tubeRack",
+        "displayVolumeUnits": "µL",
+        "tags": []
+    },
+    "dimensions": {
+        "xDimension": 127.75,
+        "yDimension": 85.5,
+        "zDimension": 77.6
+    },
+    "wells": {
+        "A1": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 18.21,
+            "y": 75.43,
+            "z": 38.5
+        },
+        "B1": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 18.21,
+            "y": 56.15,
+            "z": 38.5
+        },
+        "C1": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 18.21,
+            "y": 36.87,
+            "z": 38.5
+        },
+        "D1": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 18.21,
+            "y": 17.59,
+            "z": 38.5
+        },
+        "A2": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 38.1,
+            "y": 75.43,
+            "z": 38.5
+        },
+        "B2": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 38.1,
+            "y": 56.15,
+            "z": 38.5
+        },
+        "C2": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 38.1,
+            "y": 36.87,
+            "z": 38.5
+        },
+        "D2": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 38.1,
+            "y": 17.59,
+            "z": 38.5
+        },
+        "A3": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 57.99,
+            "y": 75.43,
+            "z": 38.5
+        },
+        "B3": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 57.99,
+            "y": 56.15,
+            "z": 38.5
+        },
+        "C3": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 57.99,
+            "y": 36.87,
+            "z": 38.5
+        },
+        "D3": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 57.99,
+            "y": 17.59,
+            "z": 38.5
+        },
+        "A4": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 77.88,
+            "y": 75.43,
+            "z": 38.5
+        },
+        "B4": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 77.88,
+            "y": 56.15,
+            "z": 38.5
+        },
+        "C4": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 77.88,
+            "y": 36.87,
+            "z": 38.5
+        },
+        "D4": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 77.88,
+            "y": 17.59,
+            "z": 38.5
+        },
+        "A5": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 97.77,
+            "y": 75.43,
+            "z": 38.5
+        },
+        "B5": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 97.77,
+            "y": 56.15,
+            "z": 38.5
+        },
+        "C5": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 97.77,
+            "y": 36.87,
+            "z": 38.5
+        },
+        "D5": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 97.77,
+            "y": 17.59,
+            "z": 38.5
+        },
+        "A6": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 117.66,
+            "y": 75.43,
+            "z": 38.5
+        },
+        "B6": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 117.66,
+            "y": 56.15,
+            "z": 38.5
+        },
+        "C6": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 117.66,
+            "y": 36.87,
+            "z": 38.5
+        },
+        "D6": {
+            "depth": 39.1,
+            "totalLiquidVolume": 1500,
+            "shape": "circular",
+            "diameter": 9.7,
+            "x": 117.66,
+            "y": 17.59,
+            "z": 38.5
+        }
+    },
+    "groups": [
+        {
+            "metadata": {
+                "wellBottomShape": "v",
+                "displayCategory": "tubeRack"
+            },
+            "wells": [
+                "A1",
+                "B1",
+                "C1",
+                "D1",
+                "A2",
+                "B2",
+                "C2",
+                "D2",
+                "A3",
+                "B3",
+                "C3",
+                "D3",
+                "A4",
+                "B4",
+                "C4",
+                "D4",
+                "A5",
+                "B5",
+                "C5",
+                "D5",
+                "A6",
+                "B6",
+                "C6",
+                "D6"
+            ]
+        }
+    ],
+    "parameters": {
+        "format": "irregular",
+        "quirks": [],
+        "isTiprack": false,
+        "isMagneticModuleCompatible": false,
+        "loadName": "E1415-1500_starlabtubes_24_tuberack_1500ul"
+    },
+    "namespace": "custom_beta",
+    "version": 1,
+    "schemaVersion": 2,
+    "cornerOffsetFromSlot": {
+        "x": 0,
+        "y": 0,
+        "z": 0
+    }
+}
